requirements.txt
python-telegram-bot>=21.0
openpyxl>=3.1.0
requests>=2.31.0
google-generativeai>=0.7.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
Pillow>=10.0.0
